Split row stream executor
