+++ jetro-core/src/io/mod.rs
@@ -18,6 +18,7 @@ mod ndjson_rev;
 mod source;
 #[allow(dead_code)]
 mod stream_plan;
+mod stream_exec;
 
 pub use ndjson::{
     collect_ndjson, collect_ndjson_file, collect_ndjson_file_with_options, collect_ndjson_matches,
